chatbot_python/requirement.txt
# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# AI e LangChain
langchain==0.0.350
langchain-community==0.0.10
google-generativeai==0.3.2
langchain-google-genai==0.0.5
langchain-core==0.1.16

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart

# document processing & embeddings
pypdf==3.17.4           
unstructured[local]==0.7.0 

# vector storage
pgvector        

# helpers
tqdm==4.66.1           